+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionId.kt
@@ -134,4 +134,7 @@ enum class ActionId {
     ANSWER_PHONE_CALL,
     END_PHONE_CALL,
     DEVICE_CONTROLS,
+
+    FORCE_STOP_APP,
+    CLEAR_RECENT_APP
 }