+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ActionUiHelper.kt
@@ -259,7 +259,7 @@ class ActionUiHelper(
                                 R.string.action_toggle_front_flashlight_with_strength,
                                 action.strengthPercent.toPercentString(),
 
-                            )
+                                )
                         } else {
                             getString(
                                 R.string.action_toggle_flashlight_with_strength,
@@ -574,6 +574,9 @@ class ActionUiHelper(
         is ActionData.HttpRequest -> action.description
 
         is ActionData.InteractUiElement -> action.description
+
+        ActionData.ClearRecentApp -> getString(R.string.action_clear_recent_app)
+        ActionData.ForceStopApp -> getString(R.string.action_force_stop_app)
     }
 
     fun getIcon(action: ActionData): ComposeIconInfo = when (action) {